HypergeoMat.jl Documentation

HypergeoMat.HypergeomPQ.BesselA — Method

BesselA(m, x, nu)

Compute the truncated Herz's type one Bessel function of a matrix argument given the eigen values of the matrix.

Arguments

m: truncation weight of the hypergeometric function, a positive integer

x: the eigen values, a vector of real or complex numbers

nu: the order parameter, real or complex number with real(nu)>-1

HypergeoMat.HypergeomPQ.BesselA — Method

BesselA(m, X, nu)

Compute the truncated Herz's type one Bessel function of a matrix argument. It is usually defined for a symmetric real matrix or a Hermitian complex matrix but arbitrary square matrices are allowed.

Arguments

m: truncation weight of the hypergeometric function, a positive integer

X: a square matrix, real or complex

nu: the order parameter, real or complex number with real(nu)>-1

HypergeoMat.HypergeomPQ.hypergeomPQ — Method

hypergeomPQ(m, a, b, x)

Compute the truncated hypergeometric function of a scalar argument.

Arguments

m: truncation weight of the summation, a positive integer

a: the "upper" parameters, a real or complex vector, possibly empty

b: the "lower" parameters, a real or complex vector, possibly empty

x: scalar, real or complex

HypergeoMat.HypergeomPQ.hypergeomPQ — Method

hypergeomPQ(m, a, b, x[, alpha])

Compute the truncated hypergeometric function of a matrix argument given the eigen values of the matrix.

Arguments

m: truncation weight of the summation, a positive integer

a: the "upper" parameters, a real or complex vector, possibly empty

b: the "lower" parameters, a real or complex vector, possibly empty

x: real or complex vector, the eigen values

alpha: the alpha parameter, a positive number; if missing, alpha=2 is used

HypergeoMat.HypergeomPQ.hypergeomPQ — Method

hypergeomPQ(m, a, b, X[, alpha])

Compute the truncated hypergeometric function of a matrix argument. The hypergeometric function is usually defined for a symmetric real matrix or a Hermitian complex matrix but arbitrary square matrices are allowed.

Arguments

m: truncation weight of the summation, a positive integer

a: the "upper" parameters, a real or complex vector, possibly empty

b: the "lower" parameters, a real or complex vector, possibly empty

X: a square matrix, real or complex

alpha: the alpha parameter, a positive number; if missing, alpha=2 is used

HypergeoMat.Mvgamma.lmvgamma — Method

lmvgamma(z, p)

Compute the logarithm of the multivariate Gamma function.

Arguments

z: real or complex number

p: positive integer, the dimension

HypergeoMat.Mvgamma.mvgamma — Method

mvgamma(z, p)

Compute the multivariate Gamma function.

Arguments

z: real or complex number

p: positive integer, the dimension